+++ client/src/views/components/inputs/UserAssignForm.jsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { useUser } from "../../../hooks/useUser";
 import { BiSend } from "react-icons/bi";
@@ -6,16 +6,18 @@ import { GiCancel } from "react-icons/gi";
 import { FaEdit } from "react-icons/fa";
 import useAxiosPrivate from "../../../hooks/useAxiosPrivate";
 import { useDevice } from "../../../hooks/useDevice";
+import { useAuth } from "../../../hooks/useAuth";
 const UserAssignForm = ({ deviceData }) => {
   const {
     register,
     handleSubmit,
     formState: { errors },
   } = useForm();
-  const { users, userLoading } = useUser();
+  const { users, userLoading, getUsers } = useUser();
   const { getDevices } = useDevice();
   const axiosPrivate = useAxiosPrivate();
   const [edit, setEdit] = useState(false);
+  const { isAuthenticated } = useAuth();
 
   const onSubmit = async (formData) => {
     try {
@@ -23,16 +25,20 @@ const UserAssignForm = ({ deviceData }) => {
         `/api/device/${deviceData.deviceId}`,
         formData
       );
-      console.log(data);
       if (data) {
         getDevices();
         setEdit(false);
       }
-      console.log(formData);
     } catch (error) {
       alert(error.message);
     }
   };
+
+  useEffect(() => {
+    if (isAuthenticated) {
+      getUsers();
+    }
+  }, [getUsers, isAuthenticated]);
 
   return (
     <>
@@ -62,7 +68,6 @@ const UserAssignForm = ({ deviceData }) => {
               <BiSend />
             </button>
           </div>
-          {console.log(errors)}
         </form>
       ) : (
         <div className="flex flex-row justify-center content-center  flex-nowrap">